ref(network): introduce consensus module
